Struct tf_rs::client::ClientSession

pub struct ClientSession<'g> { /* fields omitted */ }

A ClientSession object lets the caller drive the evaluation of the TensorFlow graph constructed with the Rust API.

Methods

impl<'g> ClientSession<'g>

pub fn new(context: &'g mut Scope) -> Result<ClientSession<'g>, Error>

pub fn reinitialize_vars(&mut self) -> &mut Self

If variables had already been initialized, reinitialize them again.

Default for this setting is false.

pub fn fetch<I, V>(&mut self, fetches: I) -> &mut Self where
I: IntoIterator<Item = V>,
V: Into<NodeIdent>,

Output to fetch from the given operation.

pub fn feed<I, Id>(&mut self, inputs: I) -> &mut Self where
Id: Into<NodeIdent>,
I: IntoIterator<Item = (Id, Vec<TensorContent>)>,

Input to feed to a graph node.

pub fn prune(&mut self) -> &mut Self

Prune the session of previous feed/fetch inputs and results.

pub fn unwrap_graph(self) -> Result<Graph, Rc<RefCell<Graph>>>

Consumes self and returns underlying graph if it's a unique reference, otherwise will return a Rc pointer to it.

pub fn run(
&mut self,
options: Option<SessionOptions>
) -> Result<Vec<TensorContent>, Error>

Run a TensorFlow session with the currently built-in graph in context. All variables are initialized beforehand.

Evaluates the tensors provided during the session construction to fetch. The number and order of outputs will match the construction phase.

Returns error if any of the session input was inadequate.