Function tf_rs::ops::constant_initializer

pub fn constant_initializer<TeS, T>(
    context: &mut Scope, 
    value: T, 
    shape: &[TeS]
) -> Result<Constant, Error> where
    T: TensorType,
    TeS: ShapeSize, 

Initializer that generates tensors with constant values.

The resulting tensor is populated with values of type dtype, as specified by arguments value following the desired shape of the new tensor (see examples below).

Args

value: All elements of the initialized variable will be set to the corresponding value in the value argument.

shape: Shape of the initializer.