Function tf_rs::ops::range

pub fn range<Ts, Tl, Td, S>(
    context: &mut Scope, 
    start: Ts, 
    limit: Tl, 
    delta: Td, 
    name: S
) -> Result<Tensor, Error> where
    Ts: TensorOps,
    Tl: TensorOps,
    Td: TensorOps,
    S: AsRef<Path>, 

Creates a sequence of numbers.

Creates a sequence of numbers that begins at start and extends by increments of delta up to but not including limit.

The dtype of the resulting tensor is inferred from the inputs unless it is provided explicitly.

Like the Python builtin range, start defaults to 0, so that range(n) = range(0, n).

For example:

# 'start' is 3
# 'limit' is 18
# 'delta' is 3
tf.range(start, limit, delta) ==> [3, 6, 9, 12, 15]

# 'start' is 3
# 'limit' is 1
# 'delta' is -0.5
tf.range(start, limit, delta) ==> [3, 2.5, 2, 1.5]

# 'limit' is 5
tf.range(limit) ==> [0, 1, 2, 3, 4]

Args: start: A 0-D Tensor (scalar). First entry in the range. limit: A 0-D Tensor (scalar). Upper limit of sequence, exclusive. delta: A 0-D Tensor (scalar). Number that increments start. dtype: The type of the elements of the resulting tensor. name: A name for the operation. Defaults to "range".

Returns: An 1-D Tensor of type dtype.