Function tf_rs::ops::reshape

pub fn reshape<Tx, Ty, S>(
    context: &mut Scope, 
    tensor: Tx, 
    shape: Ty, 
    name: S
) -> Result<Tensor, Error> where
    Tx: Into<Tensor>,
    Ty: TensorOps,
    S: AsRef<Path>, 

Reshapes a tensor.

Given tensor, this operation returns a tensor that has the same values as tensor with shape shape.

If one component of shape is the special value -1, the size of that dimension is computed so that the total size remains constant. In particular, a shape of [-1] flattens into 1-D. At most one component of shape can be -1.

If shape is 1-D or higher, then the operation returns a tensor with shape shape filled with the values of tensor. In this case, the number of elements implied by shape must be the same as the number of elements in tensor.

Args

tensor: A Tensor.

shape: A Tensor. Must be one of the following types: int32, int64. Defines the shape of the output tensor.

name: A name for the operation (optional).