Function tf_rs::ops::transpose

pub fn transpose<S, TeS, Tx>(
    context: &mut Scope, 
    a: Tx, 
    perm: Option<TeS>, 
    name: S
) -> Result<Tensor, Error> where
    Tx: Into<Tensor>,
    TeS: TensorOps,
    S: AsRef<Path>, 

Transposes a. Permutes the dimensions according to perm.

The returned tensor's dimension i will correspond to the input dimension perm[i]. If perm is not given, it is set to (n-1...0), where n is the rank of the input tensor. Hence by default, this operation performs a regular matrix transpose on 2-D input Tensors.

For example:

# 'x' is [[1 2 3]
#         [4 5 6]]
tf.transpose(x) ==> [[1 4]
                     [2 5]
                     [3 6]]

# Equivalently
tf.transpose(x, perm=[1, 0]) ==> [[1 4]
                                  [2 5]
                                  [3 6]]

# 'perm' is more useful for n-dimensional tensors, for n > 2
# 'x' is   [[[1  2  3]
#            [4  5  6]]
#           [[7  8  9]
#            [10 11 12]]]
# Take the transpose of the matrices in dimension-0
tf.transpose(x, perm=[0, 2, 1]) ==> [[[1  4]
                                      [2  5]
                                      [3  6]]

                                     [[7 10]
                                      [8 11]
                                      [9 12]]]

Args

a: A Tensor.

perm: A permutation of the dimensions of a.

name: A name for the operation (optional).

Returns

A transposed Tensor.