Function tf_rs::ops::unsorted_segment_sum

pub fn unsorted_segment_sum<Tx, Ty, Tz, S>(
    context: &mut Scope, 
    data: Tx, 
    segment_ids: Ty, 
    num_segments: Tz, 
    name: S
) -> Result<Tensor, Error> where
    Tx: Into<Tensor>,
    Ty: Into<Tensor>,
    Tz: Into<Tensor>,
    S: AsRef<Path>, 

Computes the sum along segments of a tensor.

Computes a tensor such that (output[i] = sum_{j...} data[j...] where the sum is over tuples j... such that segment_ids[j...] == i. Unlike SegmentSum, segment_ids need not be sorted and need not cover all values in the full range of valid values.

If the sum is empty for a given segment ID i, output[i] = 0.

num_segments should equal the number of distinct segment IDs.

Returns a Tensor that has the same type as data. Has same shape as data, except for the first segment_ids.rank dimensions, which are replaced with a single dimension which has size num_segments.