Struct tf_rs::ops::Scope

pub struct Scope { /* fields omitted */ }

Master context manager for building TensorFlow graphs and managing session execution.

Methods

impl Scope

pub fn new() -> Scope

pub fn variable_scope<S>(
&mut self,
name: S,
default_name: Option<S>,
reuse: Option<bool>
) -> Result<Scope, Error> where
S: AsRef<Path>,

Returns a context manager for defining ops that creates variables (layers).

If the name argument is an empty string the name will be autogenerated.

pub fn name_scope<S>(&mut self, name: S, default: Option<S>) -> Scope where
S: AsRef<Path>,

Returns a context manager for use when defining an op.

If the name argument is an empty string the name will be autogenerated.

pub fn get_variable<IS, S>(
&mut self,
dtype: Option<DataType>,
shape: Option<IS>,
name: S
) -> Result<Variable, Error> where
S: AsRef<Path>,
IS: IntoShape,

Gets an existing variable with these parameters or create a new one.

This function prefixes the name with the current variable scope and performs reuse checks.

Returns an error when creating a new variable and shape is not declared or when violating reuse during variable creation.

Reuse is set during variable_scope creation.

pub fn get_variable_with_initializer<S, T>(
&mut self,
initializer: T,
validate_shape: bool,
name: S
) -> Result<Variable, Error> where
S: AsRef<Path>,
T: Into<NodeIdent>,

Create a new_variable with the given initializer.

pub fn constant<TeS, T, S>(
&mut self,
value: &[T],
shape: &[TeS],
name: S
) -> Result<Constant, Error> where
S: AsRef<Path>,
T: TensorType,
TeS: ShapeSize,

Create a new 'constant' tensor with given values and shape.

pub fn placeholder(&mut self, dtype: DataType) -> Tensor

Inserts a placeholder for a tensor that will be always fed.

Important: This tensor will produce an error if evaluated. Its value must be fed to the client session.

pub fn control_dependencies<'a, I, T: 'a>(&mut self, control_inputs: I) -> Scope where
I: IntoIterator<Item = &'a T>,
T: GetIdent,

Returns a scope that specifies control dependencies.

pub fn clear_control_dependencies(&mut self) -> Scope

Returns a scope which ignores all previously set up control dependencies.

pub fn identity<S, Tx>(&mut self, tensor: Tx, name: S) -> Result<Tensor, Error> where
S: AsRef<Path>,
Tx: GetIdent,

Returns a copy of the variable, with the same shape and content.

pub fn set_random_seed(&mut self, value: Option<i32>)

Sets the graph-level random seed. Can only be set at root scope context.

Operations that rely on a random seed actually derive it from two seeds: the graph-level and operation-level seeds. This sets the graph-level seed.

Its interactions with operation-level seeds is as follows: 1. If neither the graph-level nor the operation seed is set: A random seed is used for this op. 2. If the graph-level seed is set, but the operation seed is not: The system deterministically picks an operation seed in conjunction with the graph-level seed so that it gets a unique random sequence. 3. If the graph-level seed is not set, but the operation seed is set: A default graph-level seed and the specified operation seed are used to determine the random sequence. 4. If both the graph-level and the operation seed are set: Both seeds are used in conjunction to determine the random sequence.

pub fn get_seed(&self, op_seed: Option<i32>) -> (Option<i32>, Option<i32>)

Returns the local seeds an operation should use given an op-specific seed.

Given operation-specific seed, op_seed, this helper function returns two seeds derived from graph-level and op-level seeds. Many random operations internally use the two seeds to allow user to change the seed globally for a graph, or for only specific operations.

pub fn unwrap_graph(self) -> Result<Graph, Rc<RefCell<Graph>>>

Consumes self and returns underlying graph if it's a unique reference, otherwise will return a Rc pointer to it.

pub fn name(&self) -> &str

Returns this scope unique name.