Module tf_rs::prelude

Reexports

pub use super::client::ClientSession;

pub use super::OperationData;

pub use super::TypedTensor;

pub use super::train;

pub use super::ops;

Structs

Constant
Constant

NodeIdent
This is a token to identify computation elements in the graph.

Scope
Master context manager for building TensorFlow graphs and managing session execution.

Status
Holds error information when communicating with back and forth with tensorflow.

TFError
The Error type.

Tensor
Tensor

Variable
Variable

Enums

Attribute
Enumerates possible operation attributes.

DataType
Type of a single tensor element.

TensorContent
An enumeration of the the different types of tensors.

Traits

DefinedShape
Methods to determine and get the shape of a tensor if it's actually defined.