Enum tf_rs::prelude::DataType

pub enum DataType {
    UnrecognizedEnumValue(u32),
    Float,
    Double,
    Int32,
    UInt8,
    Int16,
    Int8,
    String,
    Complex64,
    Int64,
    Bool,
    QInt8,
    QUInt8,
    QInt32,
    BFloat16,
    QInt16,
    QUInt16,
    UInt16,
    Complex128,
    Half,
    Resource,
}

Type of a single tensor element.

Variants
UnrecognizedEnumValue(u32)

Represents an unrecognized value.

This allows such values to come from the C API and be sent back to the C API without loss in case new values are added in the future.

Float

32-bit floating point.

Double

64-bit floating point.

Int32

32-bit signed integer.

UInt8

8-bit unsigned integer.

Int16

16-bit signed integer.

Int8

8-bit signed integer.

String

String.

Complex64

Complex number composed of two 32-bit floats.

Int64

64-bit signed integer.

Bool

Boolean.

QInt8

Quantized 8-bit signed integer.

QUInt8

Quantized 8-bit unsigned integer.

QInt32

Quantized 32-bit signed integer.

BFloat16

Float32 truncated to 16 bits. Only for cast ops.

QInt16

Quantized 16-bit signed integer.

QUInt16

Quantized 16-bit unsigned integer.

UInt16

16-bit unsigned integer.

Complex128

Complex number composed of two 64-bit floats.

Half

16-bit floating point.

Resource

TensorFlow Resource (name, container, device,...)