Struct tf_rs::prelude::Status

pub struct Status { /* fields omitted */ }

Holds error information when communicating with back and forth with tensorflow.

It either has an Code::Ok code, or otherwise an error code with an associated message.

Methods

impl Status

fn new_set(code: Code, msg: &str) -> Result<Status, NulError>

Creates a status and sets its code and message.

fn code(&self) -> Code

Returns the status's code.

fn is_ok(&self) -> bool

Returns true if the status's code is Code::Ok.

fn set(&mut self, code: Code, msg: &str) -> Result<(), NulError>

Sets the code and message.

impl Status

fn new() -> Status

Creates a status with Code::Ok and no message.