Function tf_rs::train::nn::log_softmax

pub fn log_softmax<L, S, TeS>(
    context: &mut Scope, 
    logits: L, 
    dim: TeS, 
    name: S
) -> Result<Tensor, Error> where
    L: Into<Tensor>,
    S: AsRef<Path>,
    TeS: ShapeSize, 

Computes log softmax activations.

For each batch i and class j we have logsoftmax = logits - log(reduce_sum(exp(logits), dim))

Args: logits: A non-empty Tensor. Must be one of the following types: half, float32, float64. dim: The dimension softmax would be performed on. The default is -1 which indicates the last dimension. name: A name for the operation (optional).

Returns: A Tensor. Has the same type as logits. Same shape as logits. Error if logits is empty or dim is beyond the last dimension of logits.