Struct tf_rs::train::ExponentialMovingAverage

pub struct ExponentialMovingAverage { /* fields omitted */ }

Maintains moving averages of variables by employing an exponential decay.

When training a model, it is often beneficial to maintain moving averages of the trained parameters. Evaluations that use averaged parameters sometimes produce significantly better results than the final trained values.

The apply() method adds shadow copies of trained variables and add ops that maintain a moving average of the trained variables in their shadow copies. It is used when building the training model. The ops that maintain moving averages are typically run after each training step. The average() and average_name() methods give access to the shadow variables and their names. They are useful when building an evaluation model, or when restoring a model from a checkpoint file. They help use the moving averages in place of the last trained values for evaluations.

The moving averages are computed using exponential decay. You specify the decay value when creating the ExponentialMovingAverage object. The shadow variables are initialized with the same initial values as the trained variables. When you run the ops to maintain the moving averages, each shadow variable is updated with the formula:

shadow_variable -= (1 - decay) * (shadow_variable - variable)

This is mathematically equivalent to the classic formula below, but the use of an assign_sub op (the "-=" in the formula) allows concurrent lockless updates to the variables:

shadow_variable = decay * shadow_variable + (1 - decay) * variable

Reasonable values for decay are close to 1.0, typically in the multiple-nines range: 0.999, 0.9999, etc.

Methods

impl ExponentialMovingAverage

pub fn new<Tx>(
decay: Tx,
num_updates: Option<Tensor>,
zero_debias: bool,
name: &str
) -> ExponentialMovingAverage where
Tx: Into<Tensor>,

Creates a new ExponentialMovingAverage object.

The apply() method has to be called to create shadow variables and add ops to maintain moving averages.

The optional num_updates parameter allows one to tweak the decay rate dynamically. It is typical to pass the count of training steps, usually kept in a variable that is incremented at each step, in which case the decay rate is lower at the start of training. This makes moving averages move faster. If passed, the actual decay rate used is:

min(decay, (1 + num_updates) / (10 + num_updates))

Args:

decay: The decay to use.

num_updates: Optional count of number of updates applied to variables.

zero_debias: If true, zero debias moving-averages that are initialized with tensors.

name: Optional prefix name to use for the name of ops added in apply().

pub fn apply(
&mut self,
context: &mut Scope,
var_list: &[Tensor]
) -> Result<Group, Error>

Maintains moving averages of variables.

var_list must be a list of Tensor or Tensor objects. This method creates shadow variables for all elements of var_list. Shadow variables for Tensor objects are initialized to the variable's initial value. They will be added to the GraphKeys.MOVING_AVERAGE_VARIABLES collection. For Tensor objects, the shadow variables are initialized to 0 and zero debiased (see docstring in assign_moving_average for more details).

shadow variables are created with trainable=False and added to the GraphKeys.ALL_VARIABLES collection. They will be returned by calls to tf.global_variables().

Note that apply() can be called multiple times with different lists of variables.

Args:

var_list: A list of Tensor or Tensor objects. The variables and Tensors must be of types float16, float32, or float64.

Returns:

Operation that updates the moving averages.

Errors:

TypeError: If the arguments are not all float16, float32, or float64.

ValueError: If the moving average of one of the variables is already being computed.

pub fn average(&self, var: &Tensor) -> Option<&Variable>

Returns the Tensor holding the average of var.

Args:

var: A Tensor object.

Returns:

A Tensor object or None if the moving average of var is not maintained.

pub fn average_name(&self, context: &Scope, var: &Tensor) -> String

Returns the name of the Tensor holding the average for var.

The typical scenario for ExponentialMovingAverage is to compute moving averages of variables during training, and restore the variables from the computed moving averages during evaluations.

To restore variables, you have to know the name of the shadow variables. That name and the original variable can then be passed to a Saver() object to restore the variable from the moving average value with: saver = tf.train.Saver({ema.average_name(var): var})

average_name() can be called whether or not apply() has been called.

Args:

var: A Tensor object.

Returns:

A string: The name of the variable that will be used or was used by the ExponentialMovingAverage type to hold the moving average of var.

pub fn variables_to_restore(
&self,
scope: &Scope,
moving_avg_variables: &[Variable]
) -> HashMap<String, Tensor>

Returns a map of names to Variables to restore.

If a variable has a moving average, use the moving average variable name as the restore name; otherwise, use the variable name.

Args:

moving_avg_variables: a list of variables that require to use of the moving variable name to be restored. If empty, it will default to context.moving_average_variables() + context.trainable_variables()

Returns:

A map from restore_names to variables. The restore_name can be the moving_average version of the variable name if it exist, or the original variable name.